[backend] chore: diary Controller 계층에서 모든 ID 파라미터 int -> long 타입 변경

+++ backend/src/main/java/com/example/demo/controller/DiaryController.java
@@ -34,7 +34,7 @@ public class DiaryController {
     }
 
     @DeleteMapping("/diaries/{id}") // 일기 삭제
-    public HashMap<String, String> deleteDiary(@PathVariable(required = true) int id, @RequestParam(defaultValue = "succ") String succMsg) {
+    public HashMap<String, String> deleteDiary(@PathVariable(required = true) long id, @RequestParam(defaultValue = "succ") String succMsg) {
 
         diaryService.deleteDiary(id);
 
@@ -44,7 +44,7 @@ public class DiaryController {
     }
 
     @GetMapping("/diaries/all/{userId}")    // 사용자가 속해있는 모든 그룹의 일기 조회
-    public HashMap<String, Object> requestAllTeamDiaries(@PathVariable(required = true) int userId) {
+    public HashMap<String, Object> requestAllTeamDiaries(@PathVariable(required = true) long userId) {
         List<TeamDiariesResponseDTO> data = diaryService.requestAllTeamDiaries(userId);
 
         HashMap<String, Object> result = new HashMap<>();
@@ -54,7 +54,7 @@ public class DiaryController {
     }
 
     @GetMapping("/diaries/details/{diaryId}")   // 일기 상세 내용 요청
-    public HashMap<String, Object> requestDiaryDetails(@PathVariable(required = true) int diaryId) {
+    public HashMap<String, Object> requestDiaryDetails(@PathVariable(required = true) long diaryId) {
         DiaryDetailResponseDTO data = diaryService.requestDiaryDetails(diaryId);
 
         HashMap<String, Object> result = new HashMap<>();
@@ -65,7 +65,7 @@ public class DiaryController {
 
     // 다이어리 id 요청
     @GetMapping("/diaries/findDiaryId")   // 일기 상세 내용 요청
-    public HashMap<String, Object> requestDiaryId(@RequestParam(defaultValue = "succ") String diaryTitle, String writtenDate, int writerId) {
+    public HashMap<String, Object> requestDiaryId(@RequestParam(defaultValue = "succ") String diaryTitle, String writtenDate, long writerId) {
         List<DiaryModel> data = diaryService.requestDiaryId(diaryTitle, writtenDate, writerId);
 
         HashMap<String, Object> result = new HashMap<>();
